# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "nipa crm api",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "NIPA server"
        }
    ],
    "paths": {
...
        "/v1/order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "list of orders",
                "operationId": "7f7d2d306809f6beed5cac8feeb35683",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "create order",
                "operationId": "b708b3e0cc39b78e556fc447ad9809b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_code",
                                    "orders"
                                ],
                                "properties": {
                                    "customer_code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "NIPA9988584"
                                    },
                                    "orders": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "[{count:1,product_id:1,form_id:1,data:{}}]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/v1/order/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "show order info",
                "operationId": "a2d7140d89e6242c318aeb4546289aa6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of order",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/v1/order/{id}/complete": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "show order complete info",
                "operationId": "d30c6ba2e668707414ca1111244eecf4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of order",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/v1/orderGroup/{id}/search": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "search in orders of specific order group by name",
                "operationId": "a5d12bf79ba89104d4ae93703d9c3b4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of order group",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}